refactor: 로고 클릭과 더보기 버튼 클릭 이벤트 로직 분리 및 재귀 제거
- eventHandlers.ts 내 initEvents()로 이벤트 초기화 로직 통합
- 로고 클릭 시 발생하던 이벤트 리스너 재귀 등록 버그 수정
- main.ts의 이벤트 설정 코드를 분리하여 가독성 향상
